training_source/requirements.txt
absl-py
pandas
datasets==2.19.0
torch
numpy
scikit-learn
transformers
simplification
sentencepiece
gensim
tensorflow
nltk
tomli
einops
sentence-transformers==2.4.0
peft==0.8.2
spacy_syllables
datadreamer.dev